SPACIO-71 refactor: form to create environment
