+++ src/components/form/createEnvironment/CreateEnvironmentForm.tsx
@@ -1,0 +1,200 @@
+"use client";
+import React, { useState } from "react";
+import {
+  Box,
+  Button,
+  Typography,
+  Checkbox,
+  SelectChangeEvent,
+  Alert,
+  CircularProgress,
+  Grid,
+  FormControlLabel,
+} from "@mui/material";
+import { FormDataCreateEnv } from "@/types/Environments";
+import { useRouter } from "next/navigation";
+import { PageRoutes } from "@/utils/constants/page-routes";
+import { createEnvironment } from "@/services/environmentService";
+import GeneralInfoForm from "./GeneralInfoForm";
+import RentSettingForm from "./RentSettingForm";
+import LocationTypeForm from "./LocationTypeForm";
+import ServicesAreasForm from "./ServicesAreasForm";
+import ImagesForm from "./ImagesForm";
+
+const CreateEnvironmentForm = () => {
+  const router = useRouter();
+
+  const [formData, setFormData] = useState<FormDataCreateEnv>({
+    title: "",
+    description: "",
+    location: "",
+    latitude: 0,
+    longitude: 0,
+    typePublicKey: "",
+    servicePublicKeys: [],
+    areas: [],
+    images: [],
+    equipment: "{}",
+    pricingPolicies: [],
+    discountPolicies: [],
+    weeklySchedules: [],
+    request360Tour: false,
+    capacity: 0,
+    instantBooking: false,
+    minRentalTime: 1,
+    maxRentalTime: 24,
+    rentalUnit: "Horas",
+  });
+
+  const handleCheckboxChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const { name, checked } = e.target;
+    setFormData((prev) => ({ ...prev, [name]: checked }));
+  };
+
+  const handleInputChange = (
+    e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
+  ) => {
+    const { name, value } = e.target;
+    setFormData((prev) => ({ ...prev, [name]: value }));
+  };
+
+  const handleSelectChange = (e: SelectChangeEvent) => {
+    const { name, value } = e.target;
+    setFormData((prev) => ({ ...prev, [name]: value }));
+  };
+
+  const handleServiceChange = (e: SelectChangeEvent<string[]>) => {
+    setFormData((prev) => ({
+      ...prev,
+      servicePublicKeys: e.target.value as string[],
+    }));
+  };
+
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState("");
+
+  const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    setLoading(true);
+    setError("");
+
+    if (
+      !formData.title ||
+      !formData.description ||
+      !formData.location ||
+      !formData.typePublicKey
+    ) {
+      setError("Por favor completa todos los campos obligatorios.");
+      setLoading(false);
+      return;
+    }
+    if (formData.servicePublicKeys.length === 0) {
+      setError("Selecciona al menos un servicio.");
+      setLoading(false);
+      return;
+    }
+    if (formData.areas.length === 0) {
+      setError("Selecciona al menos un área.");
+      setLoading(false);
+      return;
+    }
+    if (formData.minRentalTime > formData.maxRentalTime) {
+      setError("El tiempo mínimo de alquiler no puede ser mayor al máximo.");
+      setLoading(false);
+      return;
+    }
+
+    try {
+      await createEnvironment(formData);
+      router.push(PageRoutes.Owner_Environments);
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    } catch (err: any) {
+      setError(err.message || "Hubo un problema al enviar el formulario");
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  return (
+    <Box
+      component="form"
+      onSubmit={handleSubmit}
+      sx={{ py: 4, px: { xs: 0, md: 6, lg: 24 }, width: "100%" }}
+    >
+      <Typography variant="h5" mb={4} textAlign={"center"}>
+        Crear Nuevo Ambiente
+      </Typography>
+
+      <Grid container spacing={10}>
+        <Grid size={{ xs: 12, md: 6 }}>
+          <GeneralInfoForm handleInputChange={handleInputChange} />
+
+          <br />
+          <hr />
+          <RentSettingForm
+            formData={formData}
+            handleInputChange={handleInputChange}
+            handleCheckboxChange={handleCheckboxChange}
+            handleSelectChange={handleSelectChange}
+          />
+        </Grid>
+
+        <Grid size={{ xs: 12, md: 6 }}>
+          <hr />
+          <LocationTypeForm
+            handleSelectChange={handleSelectChange}
+            formData={formData}
+          />
+          <br />
+          <hr />
+          <ServicesAreasForm
+            formData={formData}
+            setFormData={setFormData}
+            handleServiceChange={handleServiceChange}
+          />
+          <br />
+          <hr />
+          <ImagesForm formData={formData} setFormData={setFormData} />
+          <br />
+          <FormControlLabel
+            control={
+              <Checkbox
+                name="request360Tour"
+                checked={formData.request360Tour}
+                onChange={handleCheckboxChange}
+              />
+            }
+            label="Solicitar creación de Tour Virtual 360°"
+          />
+        </Grid>
+
+        {/* Botón Submit */}
+        <Grid size={{ xs: 12 }}>
+          <Button
+            type="submit"
+            variant="contained"
+            color="primary"
+            fullWidth
+            sx={{ mt: 2 }}
+          >
+            Crear Ambiente
+          </Button>
+        </Grid>
+
+        {/* Feedback */}
+        {error && (
+          <Alert severity="error" sx={{ mt: 2 }}>
+            {error}
+          </Alert>
+        )}
+        {loading && (
+          <Box sx={{ mt: 2, display: "flex", justifyContent: "center" }}>
+            <CircularProgress />
+          </Box>
+        )}
+      </Grid>
+    </Box>
+  );
+};
+
+export default CreateEnvironmentForm;
